requirements.txt
requests==2.31.0
python-telegram-bot==20.7
solana==0.30.2
solders==0.21.0
spl-token==0.3.0
base58==2.1.1
pytz==2023.3
python-dotenv==1.0.0